fix: cron

+++ src/main/kotlin/xyz/cssxsh/mirai/spi/GroupCurfewTimer.kt
@@ -1,6 +1,6 @@
 package xyz.cssxsh.mirai.spi
 
 /**
- * 宵禁, [run] 返回指定群是否禁言
+ * 宵禁, [run] 返回指定群禁言时间（毫秒），null 不禁言
  */
-public interface GroupCurfewTimer : GroupTimerService<Int?>
+public interface GroupCurfewTimer : GroupTimerService<Long?>
